Restructure to fix install problems
- Move top level .py files into a subdir
- Install as a `console_script` instead of a `scripts`
- Don't install tests unnecessarily
- Fix up broken references
