+++ stac_validator/stac_exceptions.py
